+++ app/src/main/java/com/uof/uof_mobile/activity/SettingActivity.java
@@ -4,7 +4,7 @@ import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.uof.uof_mobile.Global;
+import com.uof.uof_mobile.other.Global;
 import com.uof.uof_mobile.R;
 
 public class SettingActivity extends AppCompatActivity {
