+++ app/src/main/java/com/uof/uof_mobile/dialog/CheckPwDialog.java
@@ -10,7 +10,7 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.widget.AppCompatTextView;
 
 import com.google.android.material.textfield.TextInputLayout;
-import com.uof.uof_mobile.Global;
+import com.uof.uof_mobile.other.Global;
 import com.uof.uof_mobile.R;
 import com.uof.uof_mobile.activity.SettingActivity;
 import com.uof.uof_mobile.manager.HttpManager;
