+++ app/src/main/java/com/uof/uof_mobile/fragment/SettingPreferenceFragment.java
@@ -14,7 +14,7 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.uof.uof_mobile.Global;
+import com.uof.uof_mobile.other.Global;
 import com.uof.uof_mobile.R;
 import com.uof.uof_mobile.activity.LoginActivity;
 import com.uof.uof_mobile.activity.SettingActivity;
