--- app/src/main/java/com/uof/uof_mobile/listview/LobbyListViewItem.java
@@ -1,30 +1,0 @@
-package com.uof.uof_mobile.listview;
-
-import org.json.JSONArray;
-
-public class LobbyListViewItem {
-    private int ordernum; // 주문번호
-    private JSONArray menulist; // ex) [{name : "productname",count : 3}]
-
-    public LobbyListViewItem(int ordernum, JSONArray menulist) {
-        this.ordernum = ordernum;
-        this.menulist = menulist;
-    }
-
-    public int getOrdernum() {
-        return ordernum;
-    }
-
-    public void setOrdernum(int ordernum) {
-        this.ordernum = ordernum;
-    }
-
-    public JSONArray getMenulist() {
-        return menulist;
-    }
-
-    public void setMenulist(JSONArray menulist) {
-        this.menulist = menulist;
-    }
-
-}
